requirements.txt
flask
youtube-transcript-api
requests
python-dotenv